+++ include/flatcc/portable/portable_basic.h
@@ -21,5 +21,7 @@
 /* These are not supported by C11 and are general platform abstractions. */
 #include "pendian.h"
 #include "punaligned.h"
+#include "pmemaccess.h"
+
 
 #endif /* PORTABLE_BASIC_H */